pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Fast API with Azure AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-core>=1.35.0",
    "azure-identity>=1.24.0",
    "azure-search-documents>=11.5.3",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "openai>=1.101.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
